Going to Work on Fixing Halloween Map Later
So, I adjusted the Halloween map because of some off geometry, and now it won't load the file properly. I don't understand why, especially since the Christmas map was redone and that loads properly, but I'm going to need to find a way to fix that.

As a side note, the stages generate a flag at runtime. You need to remove the flag object reference from the prefab, delete the flag, then save it. I need to check again to see if it saves properly without doing this, but it's good to be safe.

+++ Assets/Scripts/Level Loader/Stage.cs
@@ -9,8 +9,8 @@ public class Stage : MonoBehaviour
 
     // the bgm for the stage.
     public AudioClip bgm;
-    public int bgmClipStart;
-    public int bgmClipEnd;
+    public float bgmClipStart;
+    public float bgmClipEnd;
 
     // the skybox for the stage
     public Material skybox;
